init Assessment3

+++ Assessment2/PartIII/header/coffee.h
@@ -6,12 +6,12 @@
 class Coffee : public MenuItem
 {
 public:
-    Coffee(std::string name, int sugar, int stock);
+    Coffee(std::string name, float sugar, int stock);
     std::string getName() const override;
     int getStock() const override;
     // void delStock();
-    int getSugar() const;
-    void addSugar(int amount);
+    float getSugar() const;
+    void addSugar(float amount);
     std::string getDetails();
     void addStock(int amount);
